+++ cas-server-core-configuration/src/main/java/org/apereo/cas/configuration/model/support/couchbase/ticketregistry/CouchbaseTicketRegistryProperties.java
@@ -1,5 +1,8 @@
 package org.apereo.cas.configuration.model.support.couchbase.ticketregistry;
 
+import org.apereo.cas.configuration.model.core.util.CryptographyProperties;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
+
 /**
  * This is {@link CouchbaseTicketRegistryProperties}.
  *
@@ -18,6 +21,17 @@ public class CouchbaseTicketRegistryProperties {
 
     private String password;
     private String bucket = "default";
+
+    @NestedConfigurationProperty
+    private CryptographyProperties crypto = new CryptographyProperties();
+
+    public CryptographyProperties getCrypto() {
+        return crypto;
+    }
+
+    public void setCrypto(final CryptographyProperties crypto) {
+        this.crypto = crypto;
+    }
 
     public boolean isQueryEnabled() {
         return queryEnabled;
