+++ cas-server-core-util/build.gradle
@@ -1,11 +1,10 @@
-
 description = "Apereo CAS Core Utilities"
 dependencies {
     compile project(":cas-server-core-api-authentication")
     compile project(":cas-server-core-api-util")
-    provided project(":cas-server-core-configuration")
     compile libraries.spring
     compile libraries.springcloud
+    compile libraries.springboot
     compile libraries.shiro
     compile libraries.quartz
     compile libraries.jackson
